@@ -1,7 +1,10 @@
+import sys
+
 import numpy as np
 from scipy.sparse import csr_matrix
 
-from qpptk import IndexDB as Index
+# from qpptk import IndexDB as Index
+from qpptk import IndexTerrier as Index
 from qpptk import QueryParserText, Config
 
 logger = Config.logger
@@ -21,7 +24,7 @@ class LocalManagerRetrieval:
         self._candidates_dict = {}
 
     def _ql_score_documents(self):
-        tf_mat = self.index.get_mat_by_terms(self.qry_terms_record.keys())
+        tf_mat = self.index.get_mat_by_terms(self.qry_terms_record)
         nnz_rows = np.unique(tf_mat.nonzero()[0])
         tf_mat = tf_mat[nnz_rows, :]
         _doc_len = self.index.get_doc_len_vec(nnz_rows)
@@ -29,24 +32,25 @@ class LocalManagerRetrieval:
         _right_log = np.array([1 - _doc_len / (self.mu + _doc_len), ]).T * np.array(
             [[v.cf for v in self.qry_terms_record.values()], ]) / self.index.total_terms
         doc_scores = np.asarray(
-            np.log(_left_log + _right_log) * np.array([[self.query[q] for q in self.qry_terms_record], ]).T).squeeze()
+            np.log(_left_log + _right_log) * np.array([[self.query[q] for q in self.qry_terms_record], ]).T)
+        doc_scores = doc_scores.reshape(doc_scores.shape[0])
         return np.sort(
-            np.array(list(zip(nnz_rows, doc_scores)), dtype=[('doc_id', np.uint32), ('doc_score', np.float)]),
+            np.array(list(zip(nnz_rows, doc_scores)), dtype=[('doc_id', np.uint32), ('doc_score', float)]),
             order='doc_score')[::-1]
 
     def rm_construction(self, ql_top_k):
         """
-        Constructing the RM without smoothing
+        Constructing the RM without smoothing.
         """
-        tf_mat = self.index.get_mat_by_docs(ql_top_k['doc_id'])
-        nnz_cols = np.unique(tf_mat.nonzero()[1])
+        tf_mat = self.index.get_mat_by_docs(ql_top_k['doc_id'])  # tf matrix by documents in results
+        nnz_cols = np.unique(tf_mat.nonzero()[1])  # all the columns (terms) that appear in the docs in the result
         tf_mat = tf_mat[:, nnz_cols]
-        exp_ql = np.exp(ql_top_k['doc_score'])
+        exp_ql = np.exp(ql_top_k['doc_score'])  # transform ln(QL) -> QL scores for all the documents
         sum_exp_ql = exp_ql.sum()
         _doc_len = self.index.get_doc_len_vec(ql_top_k['doc_id'])
-        _p_w_rm = np.array([exp_ql / _doc_len, ]) * tf_mat
-        p_w_rm = np.asarray(_p_w_rm / sum_exp_ql).squeeze()
-        return np.sort(np.array(list(zip(nnz_cols, p_w_rm)), dtype=[('term_id', np.uint32), ('term_score', np.float)]),
+        p_w_rm = ((np.array([exp_ql / _doc_len, ]) * tf_mat) / sum_exp_ql).reshape(nnz_cols.shape)  # sum p(w|d)*p(d|q)
+        # p_w_rm = _p_w_rm.reshape(nnz_cols.shape)
+        return np.sort(np.array(list(zip(nnz_cols, p_w_rm)), dtype=[('term_id', np.uint32), ('term_score', float)]),
                        order='term_score')[::-1]
 
     def rank_by_kl(self, p_w_rm_top_n, rank_doc_ids=None):
@@ -64,7 +68,7 @@ class LocalManagerRetrieval:
             tf_mat = tf_mat[nnz_rows, :]
             _doc_len = _doc_len[nnz_rows]
             doc_indices = nnz_rows
-        terms_cf = self.index.get_terms_cf_vec(p_w_rm_top_n['term_id'])
+        terms_cf = self.index.get_terms_cf_vec(p_w_rm_top_n['term_id'])  # FIXME: I'm sending term_ids, but the function expects term indices
         _left = tf_mat.multiply(csr_matrix(1 / (self.mu + _doc_len)).T)
         _right = np.array([1 - _doc_len / (self.mu + _doc_len), ]).T * np.array([terms_cf, ]) / self.index.total_terms
         doc_lm = np.log(_left + _right)
@@ -79,13 +83,13 @@ class LocalManagerRetrieval:
             return True
         return False
 
-    def _generate_matching_postings(self):
-        for term in self.query:
-            term, cf, df, posting_list = self.index.get_posting_list(term)
-            if cf == 0:
-                self.oov_terms.add(term)
-                continue
-            self._candidates_dict[term] = dict(posting_list)
+    # def _generate_matching_postings(self):
+    #     for term in self.query:
+    #         term, cf, df, posting_list = self.index.get_posting_list(term)
+    #         if cf == 0:
+    #             self.oov_terms.add(term)
+    #             continue
+    #         self._candidates_dict[term] = dict(posting_list)
 
     def translate_doc_id_to_doc_no(self, result_tuple):
         return tuple(map(lambda x: (self.index.get_doc_name(x[0]), x[1]), result_tuple))
@@ -104,14 +108,14 @@ class LocalManagerRetrieval:
             logger.info(f"Query: {self.qid}; terms {self.oov_terms} are out of vocabulary")
         return False
 
-    def _init_sorted_ql_docs(self, working_set_docs=None):
+    def _init_sorted_ql_docs(self, ranking_set_docs=None):
         if self._is_empty_query():
             return tuple()
-        if working_set_docs.any():
-            doc_scores = working_set_docs.T[1]
-            doc_ids = self.index.get_doc_ids_by_name(working_set_docs.T[0])
+        if ranking_set_docs is not None:
+            doc_scores = ranking_set_docs.T[1]
+            doc_ids = self.index.get_doc_ids_by_name(ranking_set_docs.T[0])
             sorted_ql_docs = np.array(list(zip(doc_ids, doc_scores)),
-                                      dtype=[('doc_id', np.uint32), ('doc_score', np.float)])
+                                      dtype=[('doc_id', np.uint32), ('doc_score', float)])
         else:
             sorted_ql_docs = self._ql_score_documents()
         return sorted_ql_docs
@@ -121,26 +125,47 @@ class LocalManagerRetrieval:
         if self._is_empty_query():
             return tuple()
         # self._generate_matching_postings()
-        sorted_ql_scored_docs = self._ql_score_documents()
+        try:
+            sorted_ql_scored_docs = self._ql_score_documents()
+        except:
+            logger.error('*** The process crashed here!! ***')
+            print(sys.exc_info()[0])
+            print(self.qid)
+            print(self.query.items())
         return self.translate_doc_id_to_doc_no_vec(sorted_ql_scored_docs[:self.num_docs])
 
-    def run_rm_retrieval(self, re_rank_ql=False, working_set_docs=None):
+    def run_rm_retrieval(self, ranking_set_docs=None, initial_set_docs=None, _sorted_rm_terms=None):
         """
         This method will use QL to retrieve an initial set of documents, then use the set to create a RM (RM1)
         if the re_rank_ql param is True, it will use the RM to re-rank the top QL documents with the RM.
         Otherwise, it will rank all the documents in the corpus.
         """
-        sorted_ql_docs = self._init_sorted_ql_docs(working_set_docs)
-        sorted_rm_terms = self.rm_construction(sorted_ql_docs[:self.fb_docs])
-        if re_rank_ql or working_set_docs:
-            sorted_rm_scored_docs = self.rank_by_kl(sorted_rm_terms[:self.fb_terms], sorted_ql_docs[:self.num_docs])
+        ranking_docs = None
+        if _sorted_rm_terms is not None:
+            sorted_rm_terms = _sorted_rm_terms
+            if ranking_set_docs is not None:
+                ranking_docs = self._init_sorted_ql_docs(ranking_set_docs)
+        else:
+            if initial_set_docs is not None:
+                sorted_ql_docs = self._init_sorted_ql_docs(initial_set_docs)
+                if ranking_set_docs is not None and ranking_set_docs != initial_set_docs:
+                    ranking_docs = self._init_sorted_ql_docs(initial_set_docs)
+                else:
+                    ranking_docs = sorted_ql_docs
+            else:
+                sorted_ql_docs = self._init_sorted_ql_docs(ranking_set_docs)
+                ranking_docs = sorted_ql_docs
+            sorted_rm_terms = self.rm_construction(sorted_ql_docs[:self.fb_docs])
+
+        if ranking_docs is not None:  # FIXME: TODO Shit hits the fan here!!!
+            sorted_rm_scored_docs = self.rank_by_kl(sorted_rm_terms[:self.fb_terms], ranking_docs[:self.num_docs])
         else:
             sorted_rm_scored_docs = self.rank_by_kl(sorted_rm_terms[:self.fb_terms])
         return self.translate_doc_id_to_doc_no_vec(sorted_rm_scored_docs[:self.num_docs]), sorted_rm_terms
 
-    def generate_rm(self, working_set_docs):
+    def generate_rm(self, initial_set_docs):
         """
-        This method will use QL to
+        This method will generate and return a RM based on the initial_set_docs passed to it, assuming the scores are QL
         """
-        sorted_ql_docs = self._init_sorted_ql_docs(working_set_docs)
+        sorted_ql_docs = self._init_sorted_ql_docs(initial_set_docs)
         return self.rm_construction(sorted_ql_docs)
