@@ -1,19 +1,20 @@
 from .utility_functions import *
 from .config import *
 from .load_text_index import IndexText, parse_posting_list
-from .load_ciff_index import IndexCiff, parse_index_file
 from .load_db_index import IndexDB
-from .parse_queries import QueryParserText, QueryParserCiff
+from .load_terrier_index import IndexTerrier
+from .parse_queries import QueryParserText, QueryParserCiff, QueryParserJsonl
 from .retrieval_local_manager import LocalManagerRetrieval
-# from .parse_ciff_queries import QueryParserCiff, parse_ciff_queries_file
 from .pre_retrieval_predictors import LocalManagerPredictorPre
 from .post_retrieval_predictors import LocalManagerPredictorPost
-from .index_to_db import parse_index_to_db
+from .qpptk_main import parse_args, main, get_queries_object
+from .score_replacement_prediction import replace_scores_in_run_file_with_reference_scores
 
 __all__ = ['Config', 'Posting', 'TermPosting', 'TermRecord', 'TermFrequency', 'DocRecord', 'ResultPair', 'get_file_len',
-           'read_line', 'parse_posting_list', 'binary_search', 'IndexText', 'IndexCiff', 'parse_index_file', 'IndexDB',
-           'QueryParserText', 'QueryParserCiff', 'LocalManagerRetrieval', 'LocalManagerPredictorPre', 'ensure_dir',
-           'ensure_file', 'LocalManagerPredictorPost', 'read_message', 'plot_roc',
-           'transform_list_to_counts_dict', 'jaccard_similarity', 'overlap_coefficient',
-           'set_index_dump_paths', 'add_topic_to_qdf', 'msgpack_encode', 'msgpack_decode',
-           'parse_index_to_db', 'read_trec_res_file']
+           'read_line', 'parse_posting_list', 'binary_search', 'IndexText', 'IndexDB', 'IndexTerrier',
+           'QueryParserText', 'QueryParserCiff', 'QueryParserJsonl', 'LocalManagerRetrieval',
+           'LocalManagerPredictorPre', 'ensure_dir', 'ensure_file', 'LocalManagerPredictorPost', 'read_message',
+           'plot_roc', 'transform_list_to_counts_dict', 'jaccard_similarity', 'overlap_coefficient',
+           'sorensen_dice_similarity', 'calc_ndcg', 'set_index_dump_paths', 'add_topic_to_qdf', 'msgpack_encode',
+           'msgpack_decode',  'read_trec_res_file', 'parse_args', 'main', 'get_queries_object',
+           'replace_scores_in_run_file_with_reference_scores']
