@@ -1,10 +1,13 @@
 from collections import defaultdict
 
+import numpy as np
 import pandas as pd
 from syct import timer
 from lxml import etree
+import os
+from sklearn.cluster import KMeans
 
-import qpptk.CommonIndexFileFormat_pb2 as ciff
+# import qpptk.CommonIndexFileFormat_pb2 as ciff
 from qpptk import Config, read_message, DocRecord, transform_list_to_counts_dict, add_topic_to_qdf
 
 
@@ -40,10 +43,17 @@ class QueryParserCiff:
         header, queries_dict, query_records = parse_ciff_queries_file(queries_file)
         self.queries_dict = queries_dict
         self._rename_queries_to_qid(query_records)
-        filter_queries_file: list = kwargs.get('filter_queries_file', [])
-        if filter_queries_file:
+        filter_queries_file: list = kwargs.get('filter_queries_file', '')
+        drop_queries_file: str = kwargs.get('drop_queries_file', '')
+        drop_duplicate_queries: bool = kwargs.get('drop_duplicate_queries', True)  # TODO: implement this
+
+        if bool(filter_queries_file):
             self.filter_queries = QueryParserText(filter_queries_file).get_query_ids()
             self._filter_queries()
+        if bool(drop_queries_file):
+            with open(drop_queries_file, 'r') as _f:
+                self.drop_queries = _f.read().splitlines()
+            self._drop_queries()
 
     def _rename_queries_to_qid(self, query_records):
         """
@@ -56,6 +66,13 @@ class QueryParserCiff:
     def _filter_queries(self):
         filter_qids = set(self.filter_queries).intersection(self.queries_dict.keys())
         self.queries_dict = {qid: self.queries_dict.get(qid) for qid in filter_qids}
+
+    def _drop_queries(self):
+        for qid in self.drop_queries:
+            try:
+                del self.queries_dict[qid]
+            except KeyError:
+                pass
 
     def get_query(self, qid: str) -> dict:
         """
@@ -72,28 +89,54 @@ class QueryParserCiff:
         """
         return sorted(self.queries_dict.keys())
 
+    def get_queries_df(self):
+        return pd.DataFrame.from_dict(self.queries_dict, orient='index')
+
 
 class QueryParserText:
     def __init__(self, queries_txt_file, **kwargs):
         self.queries_file = queries_txt_file
         self.raw_queries_df = self._read_queries()
         self.queries_sr = self._weight_queries()
-        filter_queries_file: list = kwargs.get('filter_queries_file', [])
-        if filter_queries_file:
+        self.__queries_df = None
+        filter_queries_file: str = kwargs.get('filter_queries_file', '')
+        drop_queries_file: str = kwargs.get('drop_queries_file', '')
+        drop_duplicate_queries: bool = kwargs.get('drop_duplicate_queries', True)
+        clusters: int = kwargs.get('cluster', None)
+        if bool(filter_queries_file):
             self.filter_queries = self._read_queries(filter_queries_file).index
             self._filter_queries()
+        if bool(drop_queries_file):
+            self.drop_queries = self._read_queries(filter_queries_file).index
+            self._drop_queries()
+        if bool(clusters):
+            self.cluster_queries_by_similarity(clusters)
+        if drop_duplicate_queries:
+            duplicated_qids = self.get_duplicates_bow()
+            self._drop_queries(duplicated_qids)
 
-    def _read_queries(self, queries_file=None):
+    @property
+    def queries_df(self):
+        if self.__queries_df is None:
+            self.__queries_df = pd.DataFrame.from_dict(self.queries_sr.to_dict(), orient='index').rename_axis(
+                index='qid')
+        return self.__queries_df
+
+    def _read_queries(self, queries_file=None):  # FIXME: add an adaptation for tsv files
         _queries_file = queries_file if queries_file else self.queries_file
         with open(_queries_file, 'r') as fp:
-            queries = [line.strip().split(' ', maxsplit=1) for line in fp]
+            queries = [line.strip().split(None, maxsplit=1) for line in fp]
         return pd.DataFrame(queries, columns=['qid', 'terms']).set_index('qid')
 
     def _weight_queries(self):
-        return self.raw_queries_df.terms.str.split().apply(transform_list_to_counts_dict)
+        return self.raw_queries_df.terms.replace('', None).dropna().str.split().apply(transform_list_to_counts_dict)
 
     def _filter_queries(self):
         self.queries_sr = self.queries_sr.loc[self.filter_queries]
+
+    def _drop_queries(self, duplicated_qids=None):
+        _duplicated_qids = self.drop_queries if duplicated_qids is None else duplicated_qids
+        self.queries_sr = self.queries_sr.drop(_duplicated_qids)
 
     def get_duplicates_bow(self):
         duplicate_qids = set()
@@ -106,6 +149,13 @@ class QueryParserText:
                 duplicate_qids.update(_df.loc[_df['terms'].str.split().map(sorted).map(tuple).duplicated('last')].index)
         return duplicate_qids
 
+    # def get_duplicates_bow_new(self):
+    #     duplicate_qids = set()
+    #     for topic, qids in add_topic_to_qdf(self.raw_queries_df)[['topic', 'qid']].groupby('topic'):
+    #         _df = self.queries_df.loc[qids['qid']].dropna(axis=1, how='all')
+    #         duplicate_qids.update(_df.loc[_df.duplicated(keep='last')].index)
+    #     return duplicate_qids
+
     def get_duplicates_seq(self):
         duplicate_qids = set()
         for topic, _df in add_topic_to_qdf(self.raw_queries_df).set_index('qid').groupby('topic')['terms']:
@@ -116,7 +166,162 @@ class QueryParserText:
         return self.queries_sr.loc[qid]
 
     def get_query_ids(self) -> list:
-        return self.queries_sr.index.tolist()
+        return self.queries_sr.index.to_numpy()
+
+    def get_queries_df(self):
+        return self.queries_df
+
+    def cluster_queries_by_similarity(self, k=3):
+        """
+        Greedy and simple (not optimal) clustering of queries by topics and similarity
+        """
+        _df = add_topic_to_qdf(self.queries_df)
+        all_topics = set(_df.topic)
+        threshold = len(all_topics) // 3
+        clusters = {i: set() for i in range(k)}
+        i = 0
+        assigned_topics = set()
+        for column in _df.set_index(['topic', 'qid']):
+            _topics = set(_df.loc[_df[column].fillna(False).astype(bool)].topic.to_numpy())
+            if len(clusters[i]) <= threshold or i == k - 1:
+                clusters[i].update(_topics.difference(assigned_topics))
+            else:
+                i += 1
+                clusters[i].update(_topics.difference(assigned_topics))
+            assigned_topics.update(_topics)
+        for i, _topics in clusters.items():
+            qids = _df[['topic', 'qid']].set_index('topic').loc[_topics].sort_index()['qid']
+            self.raw_queries_df.terms.replace('', None).dropna().str.split().apply(' '.join).loc[qids].to_csv(
+                self.queries_file + f'.part-{i}', sep='\t', header=False)
+
+
+class QueryParserJsonl:
+    def __init__(self, queries_jsonl_file, terrier_index, **kwargs):
+        self.queries_file = queries_jsonl_file
+        self.terrier_index = os.path.abspath(terrier_index)
+        self.raw_queries_df = self._read_queries()
+        self.queries_sr = self._weight_queries()
+        self.__queries_df = None
+        filter_queries_file: str = kwargs.get('filter_queries_file', '')
+        drop_queries_file: str = kwargs.get('drop_queries_file', '')
+        drop_duplicate_queries: bool = kwargs.get('drop_duplicate_queries', True)
+        clusters: int = kwargs.get('cluster', None)
+        if bool(filter_queries_file):
+            self.filter_queries = self._read_queries(filter_queries_file).index
+            self._filter_queries()
+        if bool(drop_queries_file):
+            self.drop_queries = self._read_queries(filter_queries_file).index
+            self._drop_queries()
+        if bool(clusters):
+            self.cluster_queries_by_similarity(clusters)
+        if drop_duplicate_queries:
+            duplicated_qids = self.get_duplicates_bow()
+            self._drop_queries(duplicated_qids)
+
+    @property
+    def queries_df(self):
+        if self.__queries_df is None:
+            self.__queries_df = pd.DataFrame.from_dict(self.queries_sr.to_dict(), orient='index').rename_axis(
+                index='qid')
+        return self.__queries_df
+
+    def _read_queries(self, queries_file=None):  # FIXME: add an adaptation for tsv files
+        _queries_file = queries_file if queries_file else self.queries_file
+        import pyterrier as pt
+        from tira.third_party_integrations import ensure_pyterrier_is_loaded
+        ensure_pyterrier_is_loaded()
+        from jnius import cast, autoclass
+        index_properties = cast('org.terrier.structures.PropertiesIndex', pt.IndexFactory.of(self.terrier_index)).getProperties().getProperty('termpipelines')
+
+        index_properties = index_properties.split(',')
+        # stopwords automatically handled later because they are oov
+        index_properties = [i for i in index_properties if i.lower() != 'stopwords']
+        if len(index_properties) == 1:
+            print("org.terrier.terms." + index_properties[0])
+            s = autoclass("org.terrier.terms." + index_properties[0])()
+            stemmer = lambda i: s.stem(i.strip()).strip()
+        elif len(index_properties) > 1:
+            raise ValueError('Could not handle ' + str(index_properties))
+        else:
+            stemmer = lambda i: i.strip()
+
+        t = autoclass("org.terrier.indexing.tokenisation.Tokeniser").getTokeniser()
+
+        def tokenize(text):
+            text = t.getTokens(text)
+            return ' '.join([stemmer(i) for i in text])
+
+        ret = pd.read_json(_queries_file, lines=True, dtype=str).rename(columns={'query': 'terms'})[
+            ['qid', 'terms']]
+        ret['terms'] = ret['terms'].apply(tokenize)
+
+        return ret.set_index('qid')
+
+    def _weight_queries(self):
+        return self.raw_queries_df.terms.replace('', None).dropna().str.split().apply(transform_list_to_counts_dict)
+
+    def _filter_queries(self):
+        self.queries_sr = self.queries_sr.loc[self.filter_queries]
+
+    def _drop_queries(self, duplicated_qids=None):
+        _duplicated_qids = self.drop_queries if duplicated_qids is None else duplicated_qids
+        self.queries_sr = self.queries_sr.drop(_duplicated_qids)
+
+    def get_duplicates_bow(self):
+        duplicate_qids = set()
+        for topic, _df in add_topic_to_qdf(pd.DataFrame(self.queries_sr)).set_index('qid').groupby('topic')['terms']:
+            try:
+                duplicate_qids.update(_df.loc[_df.duplicated('last')].index)
+                # duplicate_qids.update(set(_df.index).difference(_df.drop_duplicates('last').index))
+            except SystemError as err:
+                _df = add_topic_to_qdf(self.raw_queries_df).set_index('topic').loc[topic].set_index('qid', drop=True)
+                duplicate_qids.update(_df.loc[_df['terms'].str.split().map(sorted).map(tuple).duplicated('last')].index)
+        return duplicate_qids
+
+    # def get_duplicates_bow_new(self):
+    #     duplicate_qids = set()
+    #     for topic, qids in add_topic_to_qdf(self.raw_queries_df)[['topic', 'qid']].groupby('topic'):
+    #         _df = self.queries_df.loc[qids['qid']].dropna(axis=1, how='all')
+    #         duplicate_qids.update(_df.loc[_df.duplicated(keep='last')].index)
+    #     return duplicate_qids
+
+    def get_duplicates_seq(self):
+        duplicate_qids = set()
+        for topic, _df in add_topic_to_qdf(self.raw_queries_df).set_index('qid').groupby('topic')['terms']:
+            duplicate_qids.update(duplicate_qids.update(_df.loc[_df.duplicated('last')].index))
+        return duplicate_qids
+
+    def get_query(self, qid: str) -> dict:
+        return self.queries_sr.loc[qid]
+
+    def get_query_ids(self) -> list:
+        return self.queries_sr.index.to_numpy()
+
+    def get_queries_df(self):
+        return self.queries_df
+
+    def cluster_queries_by_similarity(self, k=3):
+        """
+        Greedy and simple (not optimal) clustering of queries by topics and similarity
+        """
+        _df = add_topic_to_qdf(self.queries_df)
+        all_topics = set(_df.topic)
+        threshold = len(all_topics) // 3
+        clusters = {i: set() for i in range(k)}
+        i = 0
+        assigned_topics = set()
+        for column in _df.set_index(['topic', 'qid']):
+            _topics = set(_df.loc[_df[column].fillna(False).astype(bool)].topic.to_numpy())
+            if len(clusters[i]) <= threshold or i == k - 1:
+                clusters[i].update(_topics.difference(assigned_topics))
+            else:
+                i += 1
+                clusters[i].update(_topics.difference(assigned_topics))
+            assigned_topics.update(_topics)
+        for i, _topics in clusters.items():
+            qids = _df[['topic', 'qid']].set_index('topic').loc[_topics].sort_index()['qid']
+            self.raw_queries_df.terms.replace('', None).dropna().str.split().apply(' '.join).loc[qids].to_csv(
+                self.queries_file + f'.part-{i}', sep='\t', header=False)
 
 
 class QueriesXMLWriter:
